[skip ci] move everything over to generator
